fixed menu filtering, added menu refresh button, moved db menu operations to new file (was in script-universal)
