+++ src/client/css-styles/styles-menus.css
@@ -73,8 +73,8 @@
 }
 
 #search-bar-container {
-    display: flex;
-    flex-direction: column;
+    display: grid;
+    grid-template-columns: calc(100% - 80px) 80px;
     width: 100%;
     height: auto;
     margin-top: 30px;
@@ -87,6 +87,12 @@
     margin: 0px;
     padding: 0px;
     padding-left: 3px;
+}
+
+#refresh-button {
+    width: auto;
+    color: white;
+    background-color: rgb(161, 51, 51);
 }
 
 #menu {
